tests/wpt/tests/html/canvas/element/pixel-manipulation/2d.imageData.object.round.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.imageData.object.round</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.imageData.object.round</h1>
<p class="desc">ImageData.data rounds numbers with round-to-zero</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("ImageData.data rounds numbers with round-to-zero");
_addTest(function(canvas, ctx) {

  var imgdata = ctx.getImageData(0, 0, 10, 10);
  imgdata.data[0] = 0.499;
  _assertSame(imgdata.data[0], 0, "imgdata.data[\""+(0)+"\"]", "0");
  imgdata.data[0] = 0.5;
  _assertSame(imgdata.data[0], 0, "imgdata.data[\""+(0)+"\"]", "0");
  imgdata.data[0] = 0.501;
  _assertSame(imgdata.data[0], 1, "imgdata.data[\""+(0)+"\"]", "1");
  imgdata.data[0] = 1.499;
  _assertSame(imgdata.data[0], 1, "imgdata.data[\""+(0)+"\"]", "1");
  imgdata.data[0] = 1.5;
  _assertSame(imgdata.data[0], 2, "imgdata.data[\""+(0)+"\"]", "2");
  imgdata.data[0] = 1.501;
  _assertSame(imgdata.data[0], 2, "imgdata.data[\""+(0)+"\"]", "2");
  imgdata.data[0] = 2.5;
  _assertSame(imgdata.data[0], 2, "imgdata.data[\""+(0)+"\"]", "2");
  imgdata.data[0] = 3.5;
  _assertSame(imgdata.data[0], 4, "imgdata.data[\""+(0)+"\"]", "4");
  imgdata.data[0] = 252.5;
  _assertSame(imgdata.data[0], 252, "imgdata.data[\""+(0)+"\"]", "252");
  imgdata.data[0] = 253.5;
  _assertSame(imgdata.data[0], 254, "imgdata.data[\""+(0)+"\"]", "254");
  imgdata.data[0] = 254.5;
  _assertSame(imgdata.data[0], 254, "imgdata.data[\""+(0)+"\"]", "254");
  imgdata.data[0] = 256.5;
  _assertSame(imgdata.data[0], 255, "imgdata.data[\""+(0)+"\"]", "255");
  imgdata.data[0] = -0.5;
  _assertSame(imgdata.data[0], 0, "imgdata.data[\""+(0)+"\"]", "0");
  imgdata.data[0] = -1.5;
  _assertSame(imgdata.data[0], 0, "imgdata.data[\""+(0)+"\"]", "0");

});
</script>
